gateway/requirements.txt
flask==3.0.0
Werkzeug==3.0.1
PyYAML==6.0.1
